% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_spline.R
\name{apply_spline}
\alias{apply_spline}
\title{Apply the spline smoothing to the daily forecast}
\usage{
apply_spline(dataframe, smoothig_parameter)
}
\arguments{
\item{x}{An hourly dataframe calendar with date, long term seasonality and forward structure.}
}
\value{
A hourly dataframe with 57 columns to be used for hourly estimation
}
\description{
The spline is done through the function spline ...
}
