% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrage_free.R
\name{arbitrage_free_gas}
\alias{arbitrage_free_gas}
\title{Calculate Arbitrage-Free Gas Prices}
\usage{
arbitrage_free_gas(DT, DT_history, colnames_fwd)
}
\arguments{
\item{DT}{A data.table containing the current forward price data for gas.}

\item{DT_history}{A data.table containing the historical daily gas data.}

\item{colnames_fwd}{A vector of column names to exclude from \code{DT}.}
}
\value{
A data.table with arbitrage-free gas prices.
}
\description{
This function calculates arbitrage-free gas prices using forward and historical data.
}
