% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrage_free_power.R
\name{arbitrage_free_power}
\alias{arbitrage_free_power}
\title{Transform forward price in a choerent way to eliminate arbitrage}
\usage{
arbitrage_free_power(dataframe, time_series_past, colnames_forward)
}
\arguments{
\item{x}{A dataframe with forward prices.}

\item{y}{A dataframe with historic daily prices}

\item{z}{a vector with 9 column names of dataframe: year, quarter, month, yearBL, quarterBL, monthBL, yearPL, quarterPL, monthPL.}
}
\value{
A dataframe with 17 columns
}
\description{
The mean of each period should be equal to the mean price of the same period in other time-frame ...
}
