% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_detection.R
\name{break_detection}
\alias{break_detection}
\title{Detect the breaks in the spot price time series}
\usage{
break_detection(dataframe, smp_name)
}
\arguments{
\item{x}{A dataframe.}

\item{y}{Price column name.}
}
\value{
A dataframe with 4 columns date, hour, price and break period
}
\description{
The function use changepoint algorithm ...
}
