% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_detection_gas.R
\name{break_detection_gas}
\alias{break_detection_gas}
\title{Detect the breaks in the spot price time series}
\usage{
break_detection_gas(dataframe, price_name)
}
\arguments{
\item{x}{A dataframe.}

\item{y}{Price column name.}
}
\value{
A dataframe with 4 columns date, price and break period
}
\description{
The function use changepoint algorithm ...
}
