% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_holidays.R
\docType{data}
\name{calendar_holidays}
\alias{calendar_holidays}
\title{Calendar Holidays Data}
\format{
A data frame with 7,240 rows and 60 columns:
\describe{
\item{date}{Date of the holiday}
\item{holiday_GR}{Binary (1 if holiday in Greece, 0 otherwise)}
\item{holiday_HU}{Binary (1 if holiday in Hungary, 0 otherwise)}
\item{holiday_BU}{Binary (1 if holiday in Bulgaria, 0 otherwise)}
}
}
\usage{
calendar_holidays
}
\description{
It contains data for GR, HU, BU
}
\keyword{datasets}
