% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar_d.R
\name{create_calendar}
\alias{create_calendar}
\title{Creates calendar regressor to calibrate the model in the future}
\usage{
create_calendar(dataframe)
}
\arguments{
\item{x}{A daily dataframe calendar with date and holiday.}
}
\value{
A dataframe with 31 columns
}
\description{
The regrossrs are cos, sin, yday, weekend ...
}
