% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar.R
\name{create_calendar_dd}
\alias{create_calendar_dd}
\title{Create a Calendar Data Table}
\usage{
create_calendar_dd(dataframe)
}
\arguments{
\item{dataframe}{A \code{data.table} containing at least:
\itemize{
\item \code{date} (Date): The date column.
\item \code{holiday} (numeric): Binary indicator for holidays.
}}
}
\value{
A \code{data.table} with additional time-based features.
}
\description{
This function adds various time-based features to a given data.table, including seasonal indicators,
weekday dummies, and polynomial transformations of the day-of-year.
}
