% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar.R
\name{create_calendar_ddhh}
\alias{create_calendar_ddhh}
\title{Create Hourly Calendar Regressors}
\usage{
create_calendar_ddhh(dataframe)
}
\arguments{
\item{dataframe}{A daily \code{data.table} containing at least:
\itemize{
\item \code{date} (Date): The date column.
\item \code{holiday} (numeric): Binary indicator for holidays.
\item Other seasonal and categorical indicators.
}}
}
\value{
A \code{data.table} with hourly time-based features.
}
\description{
This function expands a daily calendar into an hourly format, adding hourly time-based features.
}
