% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_calendar_h.R
\name{create_calendar_h}
\alias{create_calendar_h}
\title{Creates hourly calendar regressor to calibrate the model in the future}
\usage{
create_calendar_h(dataframe)
}
\arguments{
\item{x}{A daily dataframe calendar with date and long term seasonality.}
}
\value{
A hourly dataframe with 17 columns
}
\description{
The regrossrs are cos, sin, yday, weekend ...
}
