% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_daily_gas.R
\name{detrend_daily_gas}
\alias{detrend_daily_gas}
\title{Decompose the history spot price into Macro trend and detrended}
\usage{
detrend_daily_gas(dataframe)
}
\arguments{
\item{x}{A daily dataframe with price}
}
\value{
A dataframe with 4 columns date, trade_close, hp_trend and detr_smp_day
}
\description{
The Macro trend is derived using HP filter
}
