% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detrend_daily.R
\name{detrend_dd}
\alias{detrend_dd}
\title{Detrend Daily Data Using HP Filter}
\usage{
detrend_dd(DT, value_name)
}
\arguments{
\item{DT}{A \code{data.table} containing:
\itemize{
\item \code{date}: Date of the observation.
\item \code{value_name}: The column name representing the value to be detrended (either \code{value_day} or \code{value}).
\item \code{break_group_p}: A grouping variable for detected break periods.
}}
}
\value{
A \code{data.table} with:
\itemize{
\item \code{date}: The original date.
\item \code{value_name}: The original value (either \code{value_day} or \code{value}).
\item \code{hp_trend}: The estimated trend component from the HP filter.
\item \code{detr_value}: The detrended value.
}
}
\description{
This function applies the Hodrick-Prescott (HP) filter to detrend daily data.
}
\details{
The Hodrick-Prescott filter is used to remove cyclical fluctuations and extract the long-term trend.
The smoothing parameter (\code{lambda = 8.322 * 10^7}) follows Caldana et al. (2017).
}
\examples{
DT <- data.table(date = as.Date('2020-01-01') + 0:100,
                 value_day = rnorm(101, 100, 10),
                 break_group_p = sample(1:3, 101, replace = TRUE))
result <- detrend_dd(DT, value_name = "value_day")

}
