% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier_daily.R
\name{filter_outlier_daily}
\alias{filter_outlier_daily}
\title{Filter the daily outlier from the spot time series with breaks}
\usage{
filter_outlier_daily(dataframe, smp_name)
}
\arguments{
\item{x}{A hourly dataframe with breaks.}

\item{y}{Price column name.}
}
\value{
A dataframe with 2 columns date, price
}
\description{
The outlier eliminated have 3 std dev greater than the mean ...
}
