% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier_daily_gas.R
\name{filter_outlier_daily_gas}
\alias{filter_outlier_daily_gas}
\title{Filter the daily outlier from the spot time series with breaks}
\usage{
filter_outlier_daily_gas(dataframe)
}
\arguments{
\item{x}{A hourly dataframe with breaks.}
}
\value{
A dataframe with 3 columns date, price, breaks
}
\description{
The outlier eliminated have 3 std dev greater than the mean ...
}
