% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier.R
\name{filter_outlier_dd}
\alias{filter_outlier_dd}
\title{Filter Outliers in Daily Gas Data}
\usage{
filter_outlier_dd(DT)
}
\arguments{
\item{DT}{A \code{data.table} containing at least three columns:
\itemize{
\item \code{date}: Date of the observation.
\item \code{value}: The corresponding value for each date.
\item \code{break_group_p}: A grouping factor for breaks in the time series.
}}
}
\value{
A \code{data.table} with the following columns:
\itemize{
\item \code{date}: The original date.
\item \code{value}: The original value, with outliers removed.
\item \code{out_dummy}: A binary flag indicating whether the value was an outlier (1) or not (0).
\item \code{break_group_p}: The original break group.
}
}
\description{
This function applies the Hamilton filter to a given time series data to remove cyclical components
and detect outliers based on the deviation from the estimated trend. The function calculates
the difference between the observed values and the trend, and flags values that deviate by more
than three standard deviations as outliers. The flagged outliers are then removed from the returned data.
}
\details{
The function performs the following steps:
\itemize{
\item Applies the Hamilton filter to the time series to obtain the trend component.
\item Calculates the difference between the original values and the trend.
\item Flags values that deviate from the trend by more than three standard deviations as outliers.
\item Removes the outliers and returns the modified data.table.
}
}
\examples{
# Example usage with a sample data.table `DT`
DT <- data.table(date = seq.Date(from = as.Date("2020-01-01"), by = "day", length.out = 100),
                 value = rnorm(100), break_group_p = rep(1:5, each = 20))
result <- filter_outlier_dd(DT)

}
