% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_outlier_hourly.R
\name{filter_outlier_hourly}
\alias{filter_outlier_hourly}
\title{Filter the hourly outlier from the spot time series with breaks}
\usage{
filter_outlier_hourly(dataframe, smp_name)
}
\arguments{
\item{x}{A hourly dataframe with breaks.}

\item{y}{Price column name.}
}
\value{
A dataframe with 5 columns date, hour, price and break period and ddhh
}
\description{
The outlier eliminated have 3 std dev greater than the mean ...
}
