% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{forecast_pwr}
\alias{forecast_pwr}
\title{Forecast Power Prices}
\usage{
forecast_pwr(input_forecast = LST_FOR, gas_forecast = ENV_FOR_GAS, smooth = 13)
}
\arguments{
\item{input_forecast}{A list containing required data for power forecasting:
\itemize{
\item \code{ric_spot_gas}: RIC identifier for spot gas prices.
\item \code{ric_fwd_gas}: RIC identifier for forward gas prices.
\item \code{dt_gas_fwds}: Data.table with forward market quotes for gas.
\item \code{dt_pwr_fwds}: Data.table with forward market quotes for power.
\item \code{saved_history_pwr}: Data.table with historical power prices.
\item \code{saved_history_gas_bis}: Data.table with historical gas prices.
\item \code{model_lt_pwr}: Data.table with the long-term power model.
\item \code{model_st_pwr}: Data.table with the short-term power model.
\item \code{calendar_forecast}: Data.table defining the forecast calendar.
\item \code{last_date}: The last available historical date for calibration.
}}

\item{gas_forecast}{A data.table containing the gas price forecast with columns \code{date} and \code{smooth_corrected}.}
}
\value{
A data.table with power price forecasts, including short-term and long-term components.
}
\description{
This function forecasts power prices using long-term and short-term models, forward market quotes, and gas price dependencies.
}
