% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_pipeline_basket.R
\name{fwd_pipeline}
\alias{fwd_pipeline}
\title{Forward Curve Estimation and Basket Construction Pipeline}
\usage{
fwd_pipeline(
  commodity_main,
  coef_glm,
  start_train,
  end_train,
  start_horizon,
  end_horizon
)
}
\arguments{
\item{commodity_main}{A character string specifying the main commodity (e.g., power or gas).}

\item{coef_glm}{A table with the commodoties and coefficientes and weights form the basket.}

\item{start_train}{A Date or character string representing the start date of the training period (for historical data).}

\item{end_train}{A Date or character string representing the end date of the training period.}

\item{start_horizon}{A Date or character string representing the start date of the forecast horizon.}

\item{end_horizon}{A Date or character string representing the end date of the forecast horizon.}
}
\value{
A data.table object containing the combined arbitrage-free forward curves for the main commodity and basket.
This includes a synthetic basket forward curve based on weighted power and gas forwards, as well as a
potential CO2 curve if relevant.
}
\description{
This function constructs arbitrage-free forward curves for power, gas, and potentially CO2, using historical spot prices and forecasts.
It also estimates a synthetic basket based on a GLM model, combining historical and forecast data for each commodity.
}
\details{
The pipeline includes the following steps:
\enumerate{
\item Retrieve historical spot prices for the main commodity and its basket components (power, gas, CO2).
\item Generate continuation data and clean it for further use.
\item Estimate a weighted basket using GLM coefficients.
\item Visualize historical and weighted basket data.
\item Retrieve and process forward price data for the main commodity and its basket components.
\item Apply arbitrage-free adjustments to the forward prices for power, gas, and CO2 (if applicable).
}
}
