% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_rics}
\alias{get_rics}
\title{Retrieve Historical Market Data for a Given RIC}
\usage{
get_rics(
  rics,
  from_date = Sys.Date() - (365 * 10),
  to_date = Sys.Date(),
  interval = "daily"
)
}
\arguments{
\item{rics}{A string representing the RIC (identifier) whose historical data is to be retrieved.}

\item{from_date}{A Date object or character string representing the start date for data retrieval. Defaults to 10 years before the current date.}

\item{to_date}{A Date object or character string representing the end date for data retrieval. Defaults to the current date.}

\item{interval}{A string specifying the time interval for the data (e.g., "daily"). Defaults to "daily".}
}
\value{
A \code{data.table} containing the historical data for the specified RIC. The columns included in the returned table are:
\itemize{
\item \code{TIMESTAMP}: The date of the data point.
\item \code{HIGH}: The highest price of the day.
\item \code{LOW}: The lowest price of the day.
\item \code{OPEN}: The opening price.
\item \code{CLOSE}: The closing price.
\item \code{VOLUME}: The trading volume.
\item \code{ric_column}: The RIC identifier.
}

If no data is available, the function returns a \code{data.table} with \code{NA} values.
}
\description{
This function fetches historical market data for a given RIC (identifier) over a specified date range. It uses \code{reikonapi::get_series()} to download the data and formats the results into a \code{data.table}.
}
\details{
This function:
\itemize{
\item Calls \code{reikonapi::get_series()} to fetch market data for the specified RIC.
\item Checks if the returned data is valid; if not, it creates a placeholder \code{data.table} with \code{NA} values.
\item Extracts and formats the necessary columns (\code{TIMESTAMP}, \code{HIGH}, \code{LOW}, \code{OPEN}, \code{CLOSE}, \code{VOLUME}, \code{ric_column}).
\item Converts the \code{TIMESTAMP} column to a proper date format (\code{YYYY-MM-DD}).
\item Prints the RIC being processed.
}
}
\examples{
# Example usage of get_rics function
data <- get_rics("AAPL.O", start_date = "2020-01-01", end_date = "2023-01-01")

}
