% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_rics_d}
\alias{get_rics_d}
\title{Retrieve Daily Historical Market Data for a Given RIC}
\usage{
get_rics_d(rics, from_date = Sys.Date() - (365 * 10), to_date = Sys.Date())
}
\arguments{
\item{rics}{A character string representing the RIC (identifier) for which historical data is to be retrieved.}

\item{from_date}{A Date object or character string specifying the start date for data retrieval. Defaults to 10 years before the current date.}

\item{to_date}{A Date object or character string specifying the end date for data retrieval. Defaults to the current date.}
}
\value{
A \code{data.table} containing the historical data for the specified RIC with the following columns:
\itemize{
\item \code{date} - The date of the observation (YYYY-MM-DD).
\item \code{ric} - The RIC identifier.
\item \code{value} - The closing price of the RIC.
\item \code{volume} - The trading volume.
}

If no data is available, the function returns an empty \code{data.table}.
}
\description{
Fetches historical daily market data for a given Reuters Instrument Code (RIC) over a specified date range.
It uses \code{eikonapir::get_timeseries()} to retrieve the data and formats it into a \code{data.table}.
}
\details{
This function:
\itemize{
\item Calls \code{eikonapir::get_timeseries()} to fetch market data for the specified RIC.
\item Removes unnecessary columns (\code{HIGH}, \code{LOW}, \code{OPEN}, \code{COUNT}).
\item Converts the \code{DATE} column to the \code{YYYY-MM-DD} format.
\item Orders the dataset by \code{DATE}.
\item Prints a message summarizing the data retrieval.
}
}
\examples{
\dontrun{
data <- get_rics_d("AAPL.O", from_date = "2020-01-01", to_date = "2023-01-01")
}

}
