% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{get_rics_h}
\alias{get_rics_h}
\title{Retrieve Hourly Data for a List of RICS}
\usage{
get_rics_h(
  rics,
  from_date = Sys.Date() - (365 * 10),
  to_date = Sys.Date(),
  interval = "daily",
  sleep = 0
)
}
\arguments{
\item{rics}{A string or vector of strings representing the RICs (identifiers) for which to retrieve data.}

\item{from_date}{A Date object or character string representing the start date of the data range (default is 7 years before the current date).}

\item{to_date}{A Date object or character string representing the end date of the data range (default is the current date).}

\item{interval}{A string specifying the data interval (default is \code{'daily'}).}
}
\value{
A data.table containing time-series data for each RIC across the specified date range. The data includes \code{TIMESTAMP}, \code{CLOSE}, \code{ric_column}, and \code{hour}.
}
\description{
This function retrieves hourly time-series data for a list of RICS from the Eikon API over a specified date range and interval. It aggregates data for 24 hours and returns it as a data table.
}
\details{
This function:
\itemize{
\item Loops through 24 hours to collect data for each RIC.
\item Downloads the data using the \code{reikonapi::get_series()} function.
\item Adjusts the data format and handles missing or empty data.
}
}
\examples{
# Example usage of get_rics_h function
data <- get_rics_h(rics = "RIC_ABC", from_date = "2020-01-01", to_date = "2020-12-31")

}
