% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H_calibration.R
\name{H_calibration}
\alias{H_calibration}
\title{Calibrate hourly seasonality on future calendar}
\usage{
H_calibration(dataframe, model_h, modality)
}
\arguments{
\item{x}{An hourly dataframe calendar with date, long term seasonality and forward structure.}

\item{y}{The modality of calibration: can be 'forecast' if forward are used or 'backtest' if month average spot price are used. default is forecast}
}
\value{
A hourly dataframe with 15 columns to be used for hourly estimation
}
\description{
The calibration is done using the nls esimated with hourly model function ...
}
