% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_model.R
\name{hourly_model}
\alias{hourly_model}
\title{Estimate the hourly seasonality model}
\usage{
hourly_model(dataframe, gas_name)
}
\arguments{
\item{x}{A hourly dataframe.}

\item{y}{Gas Price column name.}
}
\value{
A nls object representing the hourly model
}
\description{
The function use a nls function ...
}
