% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hourly_regressors.R
\name{hourly_regressors}
\alias{hourly_regressors}
\title{Create regressors in the hourly dataframe to estimate the model}
\usage{
hourly_regressors(dataframe)
}
\arguments{
\item{x}{A hourly dataframe with price produced by filter outlier hourly}
}
\value{
A dataframe with 49 columns
}
\description{
creates calendar variable, break dummy and bl polynomial
}
