% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_inputs}
\alias{load_inputs}
\title{Load Input Data and Parameters for HPFC Model}
\usage{
load_inputs(params, manual_data = NULL, reuters_key = NULL)
}
\arguments{
\item{params_path}{Character. Path to the JSON file containing simulation parameters.
Default is \code{'params.json'}.}
}
\value{
A list containing:
\itemize{
\item \code{LST_PARAMS} - List of parameters loaded from the JSON file.
\item \code{LST_DIRS} - List of directories for storing data.
\item \code{LST_DIRS_archive} - List of directories specific to the simulation name.
\item \code{ENV_CODES} - List containing calendar holiday information.
\item \code{ENV_SPOT} - List with historical gas and power spot prices.
\item \code{ENV_FWD} - List containing forward market data and calendar information.
}
}
\description{
This function loads simulation parameters from a JSON file and retrieves relevant historical
and forward market data for power and gas prices. It also initializes directories for storing
data and returns an environment with relevant variables.
}
