% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_regressors.R
\name{long_term_regressor}
\alias{long_term_regressor}
\title{Create regressors in the daily dataframe to estimate the model}
\usage{
long_term_regressor(dataframe, alpha)
}
\arguments{
\item{x}{A daily dataframe with date, hp_trend and detr_smp_day}

\item{y}{a parameter for weight}
}
\value{
A dataframe with 41 columns
}
\description{
The Macro trend is derived using HP filter
}
