% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_regressor_gas.R
\name{long_term_regressor_gas}
\alias{long_term_regressor_gas}
\title{Create regressors in the daily dataframe to estimate the model}
\usage{
long_term_regressor_gas(dataframe, alpha)
}
\arguments{
\item{x}{A daily dataframe with date, hp_trend, break_group_p and detr_smp_day}

\item{y}{a parameter for weight}
}
\value{
A dataframe with 41 columns
}
\description{
The Macro trend is derived using HP filter
}
