% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LT_calibration.R
\name{LT_calibration}
\alias{LT_calibration}
\title{Calibrate long term seasonality on future calendar}
\usage{
LT_calibration(dataframe, profile_matrix)
}
\arguments{
\item{x}{A dataframe with calendar regressor, ttf and power future.}

\item{y}{Long term parameters created by the function 'long term models'.}
}
\value{
A dataframe with 52 columns among whihch L_t
}
\description{
The function use changepoint algorithm ...
}
