% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_term_model.R
\name{model_long_term}
\alias{model_long_term}
\title{Estimate the long term seasonality model}
\usage{
model_long_term(dataframe, gas_name)
}
\arguments{
\item{x}{A dataframe.}

\item{y}{Gas Price column name.}
}
\value{
A dataframe with 28 columns variable representing parameters
}
\description{
The function use a stepwise regression, This is the second version with trade_close included
}
