% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_long_term_gas.R
\name{model_long_term_gas}
\alias{model_long_term_gas}
\title{Estimate the long term seasonality model}
\usage{
model_long_term_gas(dataframe)
}
\arguments{
\item{x}{A dataframe.}
}
\value{
A dataframe with 28 columns variable representing parameters
}
\description{
The function use a stepwise regression ...
}
