% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_adjusting.R
\name{period_adjusting}
\alias{period_adjusting}
\title{Adjust the Long seasonality estimated in the calendar according to the forward structure}
\usage{
period_adjusting(dataframe, last_date)
}
\arguments{
\item{x}{An hourly dataframe calendar with date, long term seasonality and forward structure.}

\item{y}{A string or date with the last date available for spot price history}
}
\value{
A hourly dataframe with 15 columns to be used for hourly estimation
}
\description{
The regrossrs are cos, sin, yday, weekend ...
}
