% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period_calibration.R
\name{period_calibration}
\alias{period_calibration}
\title{Adjust Calibrates and Detrend Data by Forward Period}
\usage{
period_calibration(DT, last_date)
}
\arguments{
\item{DT}{A data.table containing the power data with necessary columns.}

\item{last_date}{A string representing the last date in 'YYYY-MM-DD' format.}
}
\value{
A data.table with adjusted periods and detrended values for 'L_t'.
}
\description{
This function adjusts the periods based on the forward calendar and detrends the 'L_t' values for historical periods. It also categorizes dates as 'history' or 'forecast' based on a provided last date.
}
