% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PL_correction.R
\name{PL_correction}
\alias{PL_correction}
\title{Apply the PL correction to the hourly forecast}
\usage{
PL_correction(dataframe)
}
\arguments{
\item{x}{An hourly dataframe calendar with date, and the forecast adjusted.}
}
\value{
A hourly dataframe with 71 columns with the final_forecast
}
\description{
The PL correction is applied to make the price coeherent with Peak load price if available ...
}
