% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_lt_gas}
\alias{predict_lt_gas}
\title{Gas Prediction Function}
\usage{
predict_lt_gas(DT, profile_matrix)
}
\arguments{
\item{DT}{A data.table containing the input data, including relevant columns like 'yday', 'holiday', 'summer', etc.}

\item{profile_matrix}{A matrix containing the long-term and seasonal parameters to be used in the prediction.}
}
\value{
A data.table with the predicted 'L_t' values and updated column names.
}
\description{
This function predicts gas-related parameters ('L_t') using a series of long-term and seasonal parameters provided in the 'profile_matrix'.
}
