% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_lt_pwr}
\alias{predict_lt_pwr}
\title{Long-term Power Prediction}
\usage{
predict_lt_pwr(DT, profile_matrix)
}
\arguments{
\item{DT}{A data.table containing the input data, including relevant columns like 'yday', 'value_gas', 'holiday', etc.}

\item{profile_matrix}{A matrix containing the long-term and seasonal parameters to be used in the prediction.}
}
\value{
A data.table with the predicted 'L_t' values.
}
\description{
This function predicts long-term power values ('L_t') using a series of parameters provided in the 'profile_matrix'.
}
