% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_st_pwr}
\alias{predict_st_pwr}
\title{Short-Term Power Prediction}
\usage{
predict_st_pwr(DT, model_h)

predict_st_pwr(DT, model_h)
}
\arguments{
\item{DT}{A data.table containing the input data, including time-related variables, gas values, and any additional predictors.}

\item{model_h}{A pre-trained model used to predict the hourly power values.}
}
\value{
A data.table with the predicted power values (\code{L_u}, \code{L_e_u}) and relevant columns.

A data.table with the predicted power values (\code{L_u}, \code{L_e_u}) and relevant columns.
}
\description{
This function predicts short-term power parameters (\code{L_u}, \code{L_e_u}) using a pre-trained model \code{model_h} and data from \code{DT}.

This function predicts short-term power parameters (\code{L_u}, \code{L_e_u}) using a pre-trained model \code{model_h} and data from \code{DT}.
}
