% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressors.R
\name{regressors_lt_model_gas}
\alias{regressors_lt_model_gas}
\title{Create regressors in the daily DT for gas model estimation}
\usage{
regressors_lt_model_gas(DT, alpha)
}
\arguments{
\item{DT}{A \code{data.table} containing at least the following columns:
\itemize{
\item \code{date} (Date): The date of each observation.
\item \code{break_group_p} (integer): Grouping variable for breakpoints.
\item \code{holiday} (numeric): Binary indicator for holidays.
\item \code{hp_trend} (numeric): HP filter trend component.
\item \code{detr_smp_day} (numeric): The de-trended daily values.
}}

\item{alpha}{A numeric weight parameter for exponential smoothing (must be between 0 and 1).}
}
\value{
A \code{data.table} with additional regressors for modeling.
}
\description{
This function generates regressors for time series modeling, incorporating seasonal,
long-term trends, and break group interactions.
}
