% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressors.R
\name{regressors_lt_model_pwr}
\alias{regressors_lt_model_pwr}
\title{Long Term model - PWR}
\usage{
regressors_lt_model_pwr(DT, alpha)
}
\arguments{
\item{DT}{A \code{data.table} with at least the following columns:
\itemize{
\item \code{date} (Date): The date of each observation.
\item \code{hp_trend} (numeric): The HP trend component.
\item \code{detr_smp_day} (numeric): The de-trended daily values.
}}

\item{alpha}{A numeric weight parameter for exponential smoothing (must be between 0 and 1).}
}
\value{
A \code{data.table} with 41 columns containing original data and additional regressors.
}
\description{
This function generates regressors for time series modeling, incorporating seasonal and long-term trends.
The macro trend is derived using the HP filter.
}
