% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuters_months.R
\docType{data}
\name{reuters_months}
\alias{reuters_months}
\title{Reuters months conversion table}
\format{
\subsection{\code{who}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{reuters code}{character}
\item{month}{month number}
}
}
}
\usage{
reuters_months
}
\description{
It contains data for GR, HU, BU
}
\keyword{datasets}
