% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reuters_quarters_PWR.R
\docType{data}
\name{reuters_quarters_PWR}
\alias{reuters_quarters_PWR}
\title{Reuters quarters PWR conversion table}
\format{
\subsection{\code{who}}{

A data frame with 12 rows and 2 columns:
\describe{
\item{reuters code}{character}
\item{quarter}{quarter number}
}
}
}
\usage{
reuters_quarters_PWR
}
\description{
It contains data for conversion
}
\keyword{datasets}
