% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{simulate_random_walk}
\alias{simulate_random_walk}
\title{Simulate multiple random walk price paths}
\usage{
simulate_random_walk(S0, sigma, Td, N)
}
\arguments{
\item{S0}{Initial price.}

\item{sigma}{Daily volatility.}

\item{Td}{Number of time steps (days).}

\item{N}{Number of simulations to run.}
}
\value{
A \code{data.table} with \code{time_index} and simulated price columns.
}
\description{
Simulate multiple random walk price paths
}
