% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo.R
\name{simulate_single_path}
\alias{simulate_single_path}
\title{Simulate a single random walk price path}
\usage{
simulate_single_path(S0, sigma, Td, N = NULL)
}
\arguments{
\item{S0}{Initial price.}

\item{sigma}{Daily volatility.}

\item{Td}{Number of time steps (days).}

\item{N}{Optional. Simulation ID (used for naming the output column).}
}
\value{
A \code{data.table} with \code{time_index} and simulated price path.
}
\description{
Simulate a single random walk price path
}
