% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline_gas}
\alias{spline_gas}
\title{Apply Spline to Gas Data}
\usage{
spline_gas(DT, smoothig_parameter)
}
\arguments{
\item{DT}{A data.table containing the power data with the 'L_e_u' column.}

\item{smoothig_parameter}{A numeric or integer parameter to control the smoothing.}
}
\value{
A data.table with adjusted 'L_e_u' values based on the spline smoothing.
}
\description{
This function applies a spline smoothing technique to power data to adjust for daily variations using a smoothing parameter.
}
