% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{spline_pwr}
\alias{spline_pwr}
\title{Apply Spline to Power with Weekly Deviation Adjustment}
\usage{
spline_pwr(DT, smoothig_parameter = 15)
}
\arguments{
\item{DT}{A data.table containing the power data with the 'L_e_u' column.}

\item{smoothig_parameter}{A numeric or integer parameter to control the smoothing.}
}
\value{
A data.table with adjusted 'L_e_u' values, considering weekly deviation and spline smoothing.
}
\description{
This function computes the weekly deviation and applies a spline smoothing technique to adjust 'L_e_u' values, substituting with spot values before forward date.
}
