% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig_generator.R
\docType{data}
\name{spot_GAS_products_full}
\alias{spot_GAS_products_full}
\title{Spot GAS Products Data}
\format{
A data.table with columns:
\describe{
\item{products_GAS}{Gas products (e.g., 'TTF', 'PSV', etc.)}
\item{products_GAS_code}{Forward codes}
\item{spot_GAS_code}{Spot market codes}
}
}
\usage{
spot_GAS_products_full
}
\description{
This dataset contains spot gas products and their corresponding forward codes.
}
\examples{
data(spot_GAS_products_full)
}
\keyword{datasets}
