% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rig_generator.R
\docType{data}
\name{spot_PWR_products_full}
\alias{spot_PWR_products_full}
\title{Spot Power Products Data}
\format{
A data.table with columns:
\describe{
\item{countries}{Countries involved in power markets}
\item{products_PWR_code}{Forward power codes}
\item{spot_PWR_code}{Spot market codes}
}
}
\usage{
spot_PWR_products_full
}
\description{
This dataset contains spot power products and their corresponding forward codes.
}
\examples{
data(spot_PWR_products_full)
}
\keyword{datasets}
