% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{storage_pather}
\alias{storage_pather}
\title{Storage qualcosa}
\usage{
storage_pather(type = "input", folder_name, rundate, markets)
}
\arguments{
\item{type}{A daily dataframe with price}

\item{folder_name}{A daily dataframe with price}

\item{rundate}{A daily dataframe with price}

\item{markets}{A daily dataframe with price}
}
\value{
A dataframe with 4 columns date, smp_day, hp_trend and detr_smp_day
}
\description{
something
}
