% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{train_lt_gas}
\alias{train_lt_gas}
\title{Train Gas Model}
\usage{
train_lt_gas(gas_data, ric_gas)
}
\arguments{
\item{gas_data}{A data.table containing processed gas data.}
}
\value{
A data.table with trained gas model results, including computed regressors.
}
\description{
This function trains a long-term (LT) gas model by computing regressors and fitting the model.
}
