% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_lt_model_gas}
\alias{train_lt_model_gas}
\title{Train long-term gas model}
\usage{
train_lt_model_gas(DT)
}
\arguments{
\item{DT}{A \code{data.table} containing at least the following columns:
\itemize{
\item \code{detr_value} (numeric): Detrended gas price.
\item \code{break_group_p} (integer): Market regime grouping.
\item \code{cos_long_term}, \code{sin_long_term}, \code{holiday}, \code{day_2} to \code{day_7} (numeric): Regressors.
}}
}
\value{
A \code{data.table} with model coefficients, including missing terms filled with 0.
}
\description{
Fits a regression model to estimate long-term gas price effects based on break periods.
}
