% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_st_model_pwr_v1}
\alias{train_st_model_pwr_v1}
\title{Predict Short-Term model - PWR}
\usage{
train_st_model_pwr_v1(DT)
}
\arguments{
\item{DT}{A \code{data.table} containing at least the following columns:
\itemize{
\item \code{value_h} (numeric): Detrended power price.
\item \code{value_gas} (numeric): Gas price variable.
\item \code{hour_1} to \code{hour_24} (numeric): Hourly dummies.
\item \code{weekend}, \code{break_h} (numeric): Weekend and market regime indicators.
}}
}
\value{
A nonlinear least squares (NLS) model.
}
\description{
This function fits a nonlinear regression model to estimate short-term effects in power prices.
The model captures hourly patterns, seasonal effects, and interactions with gas prices.
}
