% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrage_free_gas.R
\name{arbitrage_free_gas}
\alias{arbitrage_free_gas}
\title{Transform forward price in a choerent way to eliminate arbitrage}
\usage{
arbitrage_free_gas(dataframe, time_series_past, colnames_forward)
}
\arguments{
\item{x}{A dataframe with forward prices.}

\item{y}{A dataframe with historic daily prices}

\item{z}{a vector with 6 column names of dataframe: year, quarter, month, yearBL, quarterBL, monthBL.}
}
\value{
A dataframe with 3 columns: month,year,trade_close
}
\description{
The mean of each period should be equal to the mean price of the same period in other time-frame ...
}
