% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arbitrage_free.R
\name{arbitrage_free_power}
\alias{arbitrage_free_power}
\title{Calculate Arbitrage-Free Power Prices}
\usage{
arbitrage_free_power(DT, DT_history, colnames_fwd)
}
\arguments{
\item{DT}{A data.table containing the current forward price data.}

\item{DT_history}{A data.table containing the historical daily data.}

\item{colnames_fwd}{A vector of column names to exclude from \code{DT}.}
}
\value{
A data.table with arbitrage-free power prices.
}
\description{
This function calculates arbitrage-free power prices using forward and historical data.
}
