% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar_holidays.R
\docType{data}
\name{calendar_holidays}
\alias{calendar_holidays}
\title{Calendar Holidays Data}
\format{
\subsection{\code{who}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{date}{Data}
\item{holiday_GR, holiday_HU, holiday_BU}{Country columns with binary (1 if holiday)}
}
}
}
\usage{
calendar_holidays
}
\description{
It contains data for GR, HU, BU
}
\keyword{datasets}
